Extract shared revealDelay and useScrollAndReveal hook
Deduplicate reveal delay class generation in recordings and about pages,
and consolidate the scroll listener plus IntersectionObserver ref callback
used by homepage, about, and recordings shells.
